% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_funcs.R
\name{cellranger2monocle}
\alias{cellranger2monocle}
\title{Converts and object from  extract_from_cellranger and puts it into monocle CellDataSet format.}
\usage{
cellranger2monocle(cellrext)
}
\arguments{
\item{An}{object from extract_from_cellranger}
}
\description{
Converts and object from  extract_from_cellranger and puts it into monocle CellDataSet format.
}
\keyword{cellranger}
\keyword{monocle}
