% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{color.gradient}
\alias{color.gradient}
\title{Sets the color scale for a given vector of values}
\usage{
color.gradient(x, colors = c("green", "black", "red"), colsteps = 50)
}
\arguments{
\item{x}{A vector of values}

\item{colors}{Base colors}

\item{colsteps}{The number of bins required}
}
\description{
Sets the color scale for a given vector of values
}
\keyword{color}
\keyword{scale}
