% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_funcs.R
\name{extract_from_cellranger}
\alias{extract_from_cellranger}
\title{Extracts the components need from a cellRanger aggrergated file}
\usage{
extract_from_cellranger(path, build = c("mm10", "hg38"))
}
\arguments{
\item{path}{A path to the cellranger output folder}

\item{build}{The version of the build, default is "mm10", alternative is "hg38"}
}
\description{
Extracts the components need from a cellRanger aggrergated file
}
\keyword{cellranger}
\keyword{extract}
