% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{rgl.points.hl.selected}
\alias{rgl.points.hl.selected}
\title{Plots a £d rgl plot higlighting selected cells}
\usage{
rgl.points.hl.selected(proj, ind)
}
\arguments{
\item{proj}{A 3-column matrix of coordinates}

\item{ind}{The rows requiring highlighting}
}
\description{
Plots a £d rgl plot higlighting selected cells
}
\keyword{3D}
\keyword{cells}
\keyword{highlighted}
\keyword{scatter}
