% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_funcs.R
\name{rgl_points_hl_selected}
\alias{rgl_points_hl_selected}
\title{Plots a 3d rgl plot higlighting selected cells}
\usage{
rgl_points_hl_selected(proj, ind)
}
\arguments{
\item{proj}{A 3-column matrix of coordinates}

\item{ind}{The rows requiring highlighting}
}
\description{
Plots a 3d rgl plot higlighting selected cells
}
\keyword{3D}
\keyword{cells}
\keyword{highlighted}
\keyword{scatter}
