% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep_funcs.R
\docType{class}
\name{SCDataSet-class}
\alias{SCDataSet-class}
\title{Plots a £d rgl plot higlighting selected cells}
\arguments{
\item{proj}{A 3-column matrix of coordinates}

\item{ind}{The rows requiring highlighting}
}
\description{
Plots a £d rgl plot higlighting selected cells
}
\keyword{3D}
\keyword{cells}
\keyword{highlighted}
\keyword{scatter}
